+++ src/main/kotlin/localization/Translatable.kt
@@ -1,0 +1,5 @@
+package dev.nycode.omsilauncher.localization
+
+interface Translatable {
+    val translation: Strings.() -> String
+}