+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   executable ([edfd472](https://github.com/NyCodeGHG/omsi-launcher/commit/edfd4722e238da15a9faaa11b95f0549c3e75db2))
 - Add app manifest merging logic ([ee8b73d](https://github.com/NyCodeGHG/omsi-launcher/commit/ee8b73deac980611b2e1976eb70d47a7ae45b631))
 - Add vdf parser ([d8b9ac4](https://github.com/NyCodeGHG/omsi-launcher/commit/d8b9ac4fa371fe89940e182a4c6e68421534a524))
+- Add all instance options to creation UI ([#26](https://github.com/NyCodeGHG/omsi-launcher/pull/26))
 
 ### Changed